refactor: add changes in folder structure
