+++ package.json
@@ -22,9 +22,8 @@
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.51.3",
     "react-router-dom": "^6.23.0",
-    "styled-components": "^6.1.8",
-    "zod": "^3.23.4",
-    "react-toastify": "^10.0.5"
+    "react-toastify": "^10.0.5",
+    "zod": "^3.23.4"
   },
   "devDependencies": {
     "@mui/styled-engine-sc": "^6.0.0-alpha.18",
