--- src/index.css
@@ -1,102 +1,0 @@
-:root {
-    font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
-    line-height: 1.5;
-    font-weight: 400;
-    font-size: 16px;
-
-    font-synthesis: none;
-    text-rendering: optimizeLegibility;
-}
-
-body {
-    margin: 0;
-    padding: 0;
-    display: flex;
-    place-items: center;
-    width: 100%;
-    min-height: 100vh;
-    flex-direction: column;
-    justify-content: space-between;
-}
-
-h1 {
-    font-size: 3.2em;
-    line-height: 1.1;
-}
-
-#root {
-    width: 100%;
-    position: relative;
-}
-
-.header_wrapper {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    padding: 12px 300px;
-    min-height: 60px;
-    background-color: rgba(195, 192, 192, 0.76);
-}
-
-.form_container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-}
-
-.form_wrapper {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-    margin-top: 200px;
-    width: 25%;
-}
-
-.form_message_container {
-    margin-top: 20px;
-}
-
-.form_component {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-    width: 100%;
-}
-
-.header_left {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 10px;
-}
-
-.header_right {
-    display: flex;
-    flex-direction: row;
-    gap: 10px;
-    align-items: center;
-}
-
-.header_right span {
-    cursor: pointer;
-}
-
-@keyframes logo-spin {
-    from {
-        transform: rotate(0deg);
-    }
-    to {
-        transform: rotate(360deg);
-    }
-}
-
-@media (prefers-reduced-motion: no-preference) {
-    a:nth-of-type(2) .logo {
-        animation: logo-spin infinite 20s linear;
-    }
-}
-
