+++ src/main.tsx
@@ -2,15 +2,15 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { StyledEngineProvider } from "@mui/material/styles";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import Layout from "./Layout.tsx";
+import Layout from "./modules/Layout.tsx";
 import NotFound from "./pages/NotFound.tsx";
 import Homepage from "./pages/Homepage.tsx";
 import Login from "./pages/Login.tsx";
 import Canvas from "./pages/Canvas.tsx";
-import Signup from "./pages/SIgnup.tsx";
+import Signup from "./pages/Signup.tsx";
 import ProtectedRoutes from "./ProtectedRoutes.tsx";
 import { AuthProvider } from "./context/AuthContext.tsx";
-import "./index.css";
+import "../index.css";
 
 const router = createBrowserRouter([
     {
