+++ src/modules/styles/modules.css
@@ -1,0 +1,17 @@
+.header_left {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    gap: 10px;
+}
+
+.header_right {
+    display: flex;
+    flex-direction: row;
+    gap: 10px;
+    align-items: center;
+}
+
+.header_right span {
+    cursor: pointer;
+}
