+++ src/pages/styles/pages.css
@@ -1,0 +1,37 @@
+.header_wrapper {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+    padding: 12px 300px;
+    min-height: 60px;
+    background-color: rgba(195, 192, 192, 0.76);
+}
+
+.form_container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+}
+
+.form_wrapper {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+    margin-top: 200px;
+    width: 25%;
+}
+
+.form_message_container {
+    margin-top: 20px;
+}
+
+.form_component {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+    width: 100%;
+}
